Move persist helper to avoid circular dependency
